requirements.txt
discord.py
aiohttp
python-dotenv
Pillow
matplotlib
requests
PyNaCl